sentimentanalysis/requirements.txt
openpyxl>=3.1.0
plotly==4.14.3
pandas==1.4.3
streamlit==1.12.0
Pillow==8.4.0
altair==4.1.0
